feat: Introduce SelectedFilesModel for managing selected files with drag-and-drop support
