+++ gui/file_manager_widget.py
@@ -15,235 +15,12 @@ from PyQt6.QtCore import (
 )
 from gui.pdf_filter_model import PDFFilterModel
 from PyQt6.QtGui import QStandardItemModel, QStandardItem, QIcon, QPixmap, QPainter, QFileSystemModel, QDropEvent
+from gui.selected_files_model import SelectedFilesModel
 from core.file_manager import FileManager, FileManagerError
 from utils.text_processor import TextProcessor
 from gui.styles import FileManagerStyles
 
-## CustomListView now imported from gui/custom_list_view.py
-
-## PDFFilterModel now imported from gui/pdf_filter_model.py
-
-class SelectedFilesModel(QStandardItemModel):
-    """Modelo para archivos seleccionados - solo títulos con soporte drag and drop"""
-
-    # Señal emitida cuando el orden de los archivos cambia
-    files_reordered = pyqtSignal()
-
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.selected_files: List[dict] = []  # cada dict: {'path': str, 'title': str}
-        # Variables para rastrear drag and drop
-        self.drag_source_row = -1
-        self.drag_file_path = None
-        self.actual_drop_row = -1  # Posición real del drop desde CustomListView
-
-    def supportedDropActions(self):
-        """Acciones de drop soportadas"""
-        return Qt.DropAction.MoveAction
-
-    def flags(self, index):
-        """Flags para cada item - habilitar drag and drop"""
-        default_flags = super().flags(index)
-        if index.isValid():
-            return default_flags | Qt.ItemFlag.ItemIsDragEnabled | Qt.ItemFlag.ItemIsDropEnabled
-        return default_flags | Qt.ItemFlag.ItemIsDropEnabled
-
-    def mimeData(self, indexes):
-        """Capturar datos del elemento que se está arrastrando"""
-        if indexes:
-            index = indexes[0]
-            self.drag_source_row = index.row()
-            item = self.itemFromIndex(index)
-            if item:
-                self.drag_file_path = item.data(Qt.ItemDataRole.UserRole)
-            else:
-                self.drag_file_path = None
-        return super().mimeData(indexes)
-
-    def set_actual_drop_row(self, row: int):
-        """Establecer la posición real del drop desde CustomListView"""
-        self.actual_drop_row = row
-
-    def dropMimeData(self, data, action, row, column, parent):
-        """Manejar el drop de items para reordenar manualmente"""
-        if action != Qt.DropAction.MoveAction:
-            return False
-
-        if self.drag_source_row == -1 or not self.drag_file_path:
-            return super().dropMimeData(data, action, row, column, parent)
-
-        # Determinar la posición de destino - usar la posición real si está disponible
-        target_row = self.actual_drop_row if self.actual_drop_row != -1 else row
-        if target_row == -1:
-            target_row = len(self.selected_files)
-
-        # Ajustar target_row si está moviendo hacia abajo
-        if self.drag_source_row < target_row:
-            target_row -= 1
-
-        # Hacer el movimiento manualmente
-        if self.drag_source_row != target_row and 0 <= target_row <= len(self.selected_files):
-            # Mover en la lista de archivos
-            file_path = self.selected_files.pop(self.drag_source_row)
-            self.selected_files.insert(target_row, file_path)
-
-            # Actualizar el modelo para reflejar el nuevo orden
-            self._rebuild_model()
-
-            # Emitir señal de cambio
-            self.files_reordered.emit()
-
-            # Limpiar variables de drag
-            self.drag_source_row = -1
-            self.drag_file_path = None
-            self.actual_drop_row = -1
-
-            return True
-        else:
-            self.drag_source_row = -1
-            self.drag_file_path = None
-            self.actual_drop_row = -1
-            return False
-
-    def moveRows(self, sourceParent, sourceFirst, sourceLast, destinationParent, destinationChild):
-        """Implementar moveRows para manejar el movimiento de filas correctamente"""
-        if sourceParent != destinationParent:
-            return False
-
-        # Verificar que los índices sean válidos
-        if (sourceFirst < 0 or sourceFirst >= self.rowCount() or
-            sourceLast < 0 or sourceLast >= self.rowCount() or
-            destinationChild < 0 or destinationChild > self.rowCount()):
-            return False
-
-        # Llamar a la implementación base
-        result = super().moveRows(sourceParent, sourceFirst, sourceLast, destinationParent, destinationChild)
-        return result
-
-    def removeRows(self, row, count, parent=QModelIndex()):
-        """Override removeRows para manejar eliminaciones durante drag and drop"""
-        result = super().removeRows(row, count, parent)
-        return result
-
-    def insertRows(self, row, count, parent=QModelIndex()):
-        """Override insertRows para manejar inserciones durante drag and drop"""
-        result = super().insertRows(row, count, parent)
-        return result
-
-    def _rebuild_model(self):
-        """Reconstruir el modelo basado en la lista actual de archivos"""
-        # Limpiar modelo actual
-        self.clear()
-
-        # Recrear items basado en el orden actual
-        for entry in self.selected_files:
-            title_item = QStandardItem(entry['title'])
-            title_item.setData(entry['path'], Qt.ItemDataRole.UserRole)
-            title_item.setIcon(self._get_pdf_icon())
-            title_item.setEditable(True)
-            self.appendRow(title_item)
-
-    def _sync_selected_files_from_model(self):
-        """Sincronizar la lista de archivos con el orden actual del modelo"""
-        old_files = self.selected_files.copy()
-        new_files = []
-
-        # Reconstruir la lista basada en el orden actual del modelo
-        for i in range(self.rowCount()):
-            item = self.item(i)
-            if item:
-                file_path = item.data(Qt.ItemDataRole.UserRole)
-                if file_path and file_path not in new_files:  # Evitar duplicados
-                    new_files.append(file_path)
-
-        # Solo actualizar si realmente cambió algo
-        if new_files != old_files:
-            self.selected_files = new_files
-            self.files_reordered.emit()
-
-    def add_file(self, file_path: str) -> bool:
-        """Agregar archivo a la selección"""
-        if any(entry['path'] == file_path for entry in self.selected_files):
-            return False
-        path_obj = Path(file_path)
-        title = TextProcessor.extract_title(path_obj.name)
-        entry = {'path': file_path, 'title': title}
-        self.selected_files.append(entry)
-
-        # Crear elemento solo con título
-        path_obj = Path(file_path)
-        title = TextProcessor.extract_title(path_obj.name)
-
-        # Crear item
-        title_item = QStandardItem(title)
-        title_item.setData(file_path, Qt.ItemDataRole.UserRole)
-        title_item.setIcon(self._get_pdf_icon())
-
-        # Agregar item
-        self.appendRow(title_item)
-        return True
-
-    def remove_file(self, row: int) -> bool:
-        """Remover archivo por índice de fila"""
-        if 0 <= row < len(self.selected_files):
-            self.selected_files.pop(row)
-            self.removeRow(row)
-            return True
-        return False
-
-    def move_file(self, from_row: int, to_row: int) -> bool:
-        """Mover archivo de una posición a otra"""
-        if (0 <= from_row < len(self.selected_files) and
-            0 <= to_row < len(self.selected_files) and
-            from_row != to_row):
-            # Guarda todos los títulos editados antes de mover
-            for i in range(self.rowCount()):
-                item = self.item(i)
-                if item:
-                    self.selected_files[i]['title'] = item.text()
-            entry = self.selected_files.pop(from_row)
-            self.selected_files.insert(to_row, entry)
-            self._rebuild_model()
-            return True
-        return False
-
-    def clear_files(self):
-        """Limpiar todos los archivos seleccionados"""
-        self.selected_files.clear()
-        self.clear()
-
-    def get_selected_files(self) -> List[str]:
-        """Obtener lista de archivos seleccionados"""
-        return [entry['path'] for entry in self.selected_files]
-
-    def _get_pdf_icon(self) -> QIcon:
-        """Crear icono para archivos PDF"""
-        pixmap = QPixmap(16, 16)
-        pixmap.fill(Qt.GlobalColor.transparent)
-
-        painter = QPainter(pixmap)
-        painter.setRenderHint(QPainter.RenderHint.Antialiasing)
-        painter.setBrush(Qt.GlobalColor.red)
-        painter.setPen(Qt.GlobalColor.darkRed)
-        painter.drawRoundedRect(2, 2, 12, 12, 2, 2)
-        painter.setPen(Qt.GlobalColor.white)
-        painter.drawText(4, 12, "PDF")
-        painter.end()
-
-        return QIcon(pixmap)
-
-    def setData(self, index, value, role=Qt.ItemDataRole.EditRole):
-        if role == Qt.ItemDataRole.EditRole and index.isValid():
-            row = index.row()
-            self.selected_files[row]['title'] = value
-        return super().setData(index, value, role)
-
-    def get_titles(self) -> List[str]:
-        for i in range(self.rowCount()):
-            item = self.item(i)
-            if item:
-                self.selected_files[i]['title'] = item.text()
-        return [entry['title'] for entry in self.selected_files]
+## SelectedFilesModel now imported from gui/selected_files_model.py
 
 class FileManagerWidget(QWidget):
     def get_selected_titles(self) -> List[str]:
